requirements.txt
opencv-contrib-python
tensorflow
Pillow